go/74.search-a-2d-matrix.go
package main

import "sort"

/*
 * @lc app=leetcode id=74 lang=golang
 *
 * [74] Search a 2D Matrix
 *
 * https://leetcode.com/problems/search-a-2d-matrix/description/
 *
 * algorithms
 * Medium (35.53%)
 * Total Accepted:    393.7K
 * Total Submissions: 1.1M
 * Testcase Example:  '[[1,3,5,7],[10,11,16,20],[23,30,34,50]]\n3'
 *
 * Write an efficient algorithm that searches for a value in an m x n matrix.
 * This matrix has the following properties:
 *
 *
 * Integers in each row are sorted from left to right.
 * The first integer of each row is greater than the last integer of the
 * previous row.
 *
 *
 *
 * Example 1:
 *
 *
 * Input: matrix = [[1,3,5,7],[10,11,16,20],[23,30,34,50]], target = 3
 * Output: true
 *
 *
 * Example 2:
 *
 *
 * Input: matrix = [[1,3,5,7],[10,11,16,20],[23,30,34,50]], target = 13
 * Output: false
 *
 *
 * Example 3:
 *
 *
 * Input: matrix = [], target = 0
 * Output: false
 *
 *
 *
 * Constraints:
 *
 *
 * m == matrix.length
 * n == matrix[i].length
 * 0 <= m, n <= 100
 * -10^4 <= matrix[i][j], target <= 10^4
 *
 *
 */

func searchMatrix(mat [][]int, t int) bool {
	m := len(mat)
	if m == 0 {
		return false
	}

	n := len(mat[0])
	if n == 0 {
		return false
	}

	y := sort.Search(m, func(i int) bool {
		return mat[i][0] >= t
	})

	if y < 0 {
		return false
	}
	if y < m && mat[y][0] == t {
		return true
	}

	if y == 0 {
		return mat[0][0] == t
	}

	y--
	x := sort.SearchInts(mat[y], t)
	return !(x < 0 || x >= n) && mat[y][x] == t
}
